+++ packages/headless/src/app/thunk-extra-arguments.ts
@@ -2,7 +2,7 @@ import {AnalyticsClientSendEventHook} from 'coveo.analytics';
 import {Logger} from 'pino';
 import {PreprocessRequest} from '../api/preprocess-request';
 import {validatePayloadAndThrow} from '../utils/validate-payload';
-import {EngineConfiguration} from './engine-configuration-options';
+import {EngineConfiguration} from './engine-configuration';
 import {NoopPreprocessRequest} from '../api/preprocess-request';
 
 export interface ThunkExtraArguments {
