+++ packages/headless/src/integration-tests/search-app.test.ts
@@ -1,5 +1,5 @@
 import {buildSearchEngine} from '../app/search-engine/search-engine';
-import {getSampleSearchEngineConfiguration} from '../app/search-engine/search-engine-configuration-options';
+import {getSampleSearchEngineConfiguration} from '../app/search-engine/search-engine-configuration';
 import {
   buildDateSortCriterion,
   buildRelevanceSortCriterion,
